Updated script.
Extract_control.py now parses a setup.ini, and generates a cached
version (stored in setup.shelf) which future invocations can load
for a speedup.

It has also been turned into a proper command-line utility. There
are options for it to generate a list of all available packages,
a 'control' file, an 'install' file, or 'source' files that can
be used for dowload control.